Restructure the basic document
The idea is to make it easier to maintain the spec in the future;
I've modularized some things likely to change quickly.

Further, some basic cleanup has been done in order to use asciidoctor
to produce a better looking document over time.  Some day, it may be
necessary to convert to LaTeX, but today is not that day.

At the same time, I've copied the original source material to the
file original-notes.adoc in order not to lose them inadvertently.
As further content gets added from these original notes, the file
will eventually become empty and no longer be needed.

Signed-off-by: Al Stone <ahs3@redhat.com>