Start filling in some structure and content for the initial draft
This change adds in definitions of basic use cases, and starts
building a basic structure for the document.

Additionally, a basic outline has been created for the general
purpose computing use case.  It is not meant to be complete, but
it does try to capture most of the topics that we will need to
standardize over time.

Some cosmetic changes were also made, just to make it look pretty.

Signed-off-by: Al Stone <ahs3@redhat.com>