Separate out platform sections from original document
The very first draft of the spec contained three sections: User Level
Platform, Supervisor Level Platform and Machine Level Platform.  This
change moves those to individual files; note that there was no content
for machine level, so there was no need for an additional file.

This is a structural change only.  All content remains the same.

It is unclear where the content in these files will eventually end up.
The outline of the spec is certain to change and how the content here
gets used is still to be determined.

Signed-off-by: Al Stone <ahs3@redhat.com>