Separate out the profile section from original document
The very first draft of the spec contained a section called "Profiles."
This change moves that content to an individual file.

This is a structural change only.  All content remains the same.

It is unclear where the content in this file will eventually end up.
The outline of the spec is certain to change and how the content here
gets used is still to be determined.

Signed-off-by: Al Stone <ahs3@redhat.com>