riscv-platform-spec: Initial server firmware requirements
This specifies mandatory requirements for server class platforms in
addition to the requirements in base specification.

Many dependencies and granular details are still TBD. But this is a
good starting point to understand and refine the requirements for
server platforms.

Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>

Reviewed-by: Abner Chang <renba.chang@gmail.com>

Changes in v5:
 - Moved SMBIOS under discovery section.
 - Added note from Abner for type 44 strcture.
 - Removed RV128 reference.
Changes in v4:
 - Used "required or depricated" convention.
Changes in v3:
 - Referred to the PCI chapter in UEFI specification.
Changes in v2:
 - Aligned to 80 characters.
 - Removed protocols related to graphics support.
 - Referred to SMBIOS conformance guidelines.