base: Add SW/ABI requirement section
There are several specifications, that are required for the
interoperability of SW components in the RISC-V ecosystem.
Among the most important ones are the procedure call standard,
the ELF specification, and the DWARF specification.

To avoid fragmentation and promote interoperability in the field,
let's mandate compliance to these three documents (if applicable).

Signed-off-by: Christoph Muellner <cmuellner@linux.com>